tests: add author search related tests
[WIP]
This is an effort to automate the author search test and avoid having
regressions when changing inspire-query-parser functionality.

The idea is to have a dataset for each search category we're trying to
improve searching on. That's the reason for having a separate
`search_api_client` fixture, which spins up an application with the
`FIXTURES_FILE` that was defined in each search test module.

Signed-off-by: Chris Aslanoglou <chris.aslanoglou@gmail.com>